requirements.txt
asgiref==3.4.1
astroid==2.9.0
boto3==1.20.15
botocore==1.23.15
dj-database-url==0.5.0
Django==3.2.9
django-annoying==0.10.6
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pylint==2.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
